requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.0
lxml==4.9.3